Graphics.Rendering.Chart

Description

A framework for creating 2D charts in Haskell.

For the simplest API, see the Graphics.Rendering.Chart.Easy module.

When more control is required, understanding the various data types is necessary. The basic model is that you define a value representing a chart to be displayed (eg. a Layout), and then convert it to a Renderable by applying toRenderable. This Renderable is then actually output by calling a function in an appropriate graphics backend, eg renderableToFile.

Currently, there are three types of charts:

Layout is a standard XY chart

LayoutLR is an XY chart with independent left and right axes

PieLayout is a pie chart

Layout and LayoutLR charts can be stacked vertically using the StackedLayouts type.

Renderables can be composed in arbitrary ways using the Graphics.Rendering.Chart.Grid module.

Many of the record structure involved in the API have a large number of fields. Lenses are provided to access each field. Also, for each record type, there is generally a default value, which can be accessed through the def value of the Default typeclass.

Documentation

module Graphics.Rendering.Chart.Geometry

module Graphics.Rendering.Chart.Drawing

module Graphics.Rendering.Chart.Renderable

module Graphics.Rendering.Chart.Layout

module Graphics.Rendering.Chart.Axis

module Graphics.Rendering.Chart.Plot

module Graphics.Rendering.Chart.Legend

module Graphics.Rendering.Chart.Backend.Types