Graphics.Rendering.Chart.Axis.Indexed

Description

Calculate and render indexed axes

Documentation

newtype PlotIndex

Type for capturing values plotted by index number (ie position in a list) rather than a numerical value.

Constructors

PlotIndex

Fields

plotindex_i :: Int

autoIndexAxis' :: Integral i => Bool -> [String] -> AxisFn i

Create an axis for values indexed by position. The list of strings are the labels to be used.

autoIndexAxis :: Integral i => [String] -> AxisFn i

addIndexes :: [a] -> [(PlotIndex, a)]

Augment a list of values with index numbers for plotting.