Graphics.Rendering.Chart.Axis.LocalTime

Description

Deprecated: Use Graphics.Rendering.Chart.Axis.Time module

Calculate and render time axes

Documentation

type TimeSeq = LocalTime -> ([LocalTime], [LocalTime])

TimeSeq is a (potentially infinite) set of times. When passed a reference time, the function returns a a pair of lists. The first contains all times in the set less than the reference time in decreasing order. The second contains all times in the set greater than or equal to the reference time, in increasing order.

type TimeLabelFn = LocalTime -> String

How to display a time

data TimeLabelAlignment

Constructors

UnderTicks

BetweenTicks

timeAxis

Arguments

:: TimeSeq
Set the minor ticks, and the final range will be aligned to its elements.

-> TimeSeq
Set the labels and grid.

-> TimeLabelFn

-> TimeLabelAlignment

-> TimeSeq
Set the second line of labels.

-> TimeLabelFn
Format LocalTime for labels.

-> TimeLabelAlignment

-> AxisFn LocalTime

Create an AxisFn to for a time axis.

The values to be plotted against this axis can be created with doubleFromLocalTime.

autoTimeAxis :: AxisFn LocalTime

Automatically choose a suitable time axis, based upon the time range of data. The values to be plotted against this axis can be created with doubleFromLocalTime.

days :: TimeSeq

A TimeSeq for calendar days.

months :: TimeSeq

A TimeSeq for calendar months.

years :: TimeSeq

A TimeSeq for calendar years.

Utilities

doubleFromLocalTime :: LocalTime -> Double

Map a LocalTime value to a plot coordinate.