Graphics.Rendering.Chart.Axis.Time

Description

Calculate and render time axes

Documentation

type TimeSeq = UTCTime -> ([UTCTime], [UTCTime])

TimeSeq is a (potentially infinite) set of times. When passed a reference time, the function returns a a pair of lists. The first contains all times in the set less than the reference time in decreasing order. The second contains all times in the set greater than or equal to the reference time, in increasing order.

type TimeLabelFn = UTCTime -> String

How to display a time

data TimeLabelAlignment

Constructors

UnderTicks

BetweenTicks

class TimeValue t where

A typeclass abstracting the functions we need to be able to plot against an axis of time type d.

Minimal complete definition

utctimeFromTV, tvFromUTCTime

Methods

utctimeFromTV :: t -> UTCTime

tvFromUTCTime :: UTCTime -> t

doubleFromTimeValue :: t -> Double

timeValueFromDouble :: Double -> t

timeValueAxis

Arguments

:: TimeValue t

=> TimeSeq
Set the minor ticks, and the final range will be aligned to its elements.

-> TimeSeq
Set the labels and grid.

-> TimeLabelFn

-> TimeLabelAlignment

-> TimeSeq
Set the second line of labels.

-> TimeLabelFn
Format t for labels.

-> TimeLabelAlignment

-> AxisFn t

Create an AxisFn to for a time axis.

The values to be plotted against this axis can be created with doubleFromLocalTime.

Implementation detail: PlotValue constraint is needed to use vmap.

autoTimeValueAxis :: TimeValue t => AxisFn t

Automatically choose a suitable time axis, based upon the time range of data. The values to be plotted against this axis can be created with doubleFromTimeValue.

days :: TimeSeq

A TimeSeq for calendar days.

months :: TimeSeq

A TimeSeq for calendar months.

years :: TimeSeq

A TimeSeq for calendar years.

Orphan instances

PlotValue LocalTime

Instance details

Methods

toValue :: LocalTime -> Double

fromValue :: Double -> LocalTime

autoAxis :: AxisFn LocalTime

PlotValue UTCTime

Instance details

Methods

toValue :: UTCTime -> Double

fromValue :: Double -> UTCTime

autoAxis :: AxisFn UTCTime

PlotValue Day

Instance details

Methods

toValue :: Day -> Double

fromValue :: Double -> Day

autoAxis :: AxisFn Day