Graphics.Rendering.Chart.Axis.Types

Description

Type definitions for Axes

Documentation

data AxisData x

The basic data associated with an axis showing values of type x.

Constructors

AxisData

Fields

_axis_visibility :: AxisVisibility

Which parts of the axis shall be displayed.

_axis_viewport :: Range -> x -> Double

The _axis_viewport function maps values into device coordinates.

_axis_tropweiv :: Range -> Double -> x

The _axis_tropweiv function maps device coordinates back to values.

_axis_ticks :: [(x, Double)]

The tick marks on the axis as pairs. The first element is the position on the axis (in viewport units) and the second element is the length of the tick in output coordinates. The tick starts on the axis, and positive numbers are drawn towards the plot area.

_axis_labels :: [[(x, String)]]

The labels on an axis as pairs. The first element of the pair is the position on the axis (in viewport units) and the second is the label text string. Note that multiple sets of labels can be specified, and are shown successively further away from the axis line.

_axis_grid :: [x]

The positions on the axis (in viewport units) where we want to show grid lines.

data AxisVisibility

Configures whick visual elements of a axis are shown at the appropriate edge of a plot area.

Constructors

AxisVisibility

Fields

_axis_show_line :: Bool

Whether to display a line along the axis.

_axis_show_ticks :: Bool

Whether to display the tick marks.

_axis_show_labels :: Bool

Whether to display the labels.

data AxisT x

Collect the information we need to render an axis. The bool is true if the axis direction is reversed.

Constructors

AxisT RectEdge AxisStyle Bool (AxisData x)

data AxisStyle

Control values for how an axis gets displayed.

Constructors

AxisStyle

Fields

_axis_line_style :: LineStyle

LineStyle to use for axis line and ticks.

_axis_label_style :: FontStyle

FontStyle to use for axis labels.

_axis_grid_style :: LineStyle

LineStyle to use for axis grid.

_axis_label_gap :: Double

How far the labels are to be drawn from the axis.

class Ord a => PlotValue a where

A typeclass abstracting the functions we need to be able to plot against an axis of type a

Methods

toValue :: a -> Double

fromValue :: Double -> a

autoAxis :: AxisFn a

type AxisFn x = [x] -> AxisData x

A function to generate the axis data, given the data values to be plotted against it.

defaultAxisLineStyle :: LineStyle

The default LineStyle of an axis.

defaultGridLineStyle :: LineStyle

The default LineStyle of a plot area grid.

makeAxis :: PlotValue x => ([x] -> [String]) -> ([x], [x], [x]) -> AxisData x

Construct an axis given the positions for ticks, grid lines, and labels, and the labelling function

makeAxis' :: Ord x => (x -> Double) -> (Double -> x) -> ([x] -> [String]) -> ([x], [x], [x]) -> AxisData x

Construct an axis given the positions for ticks, grid lines, and labels, and the positioning and labelling functions

axisToRenderable :: AxisT x -> Renderable x

Construct a renderable from an axis, in order that it can be composed with other renderables and drawn. This does not include the drawing of the grid, which must be done separately by the renderAxisGrid function.

renderAxisGrid :: RectSize -> AxisT z -> BackendProgram ()

axisOverhang :: Ord x => AxisT x -> BackendProgram (Double, Double)

Calculate the amount by which the labels extend beyond the ends of the axis.

vmap :: PlotValue x => (x, x) -> Range -> x -> Double

A linear mapping of points in one range to another.

invmap :: PlotValue x => (x, x) -> Range -> Double -> x

The inverse mapping from device co-ordinate range back to interesting values.

linMap :: (a -> Double) -> (a, a) -> Range -> a -> Double

A linear mapping of points in one range to another.

invLinMap :: (Double -> a) -> (a -> Double) -> (a, a) -> Range -> Double -> a

An inverse linear mapping of points from one range to another.

axisGridAtTicks :: AxisData x -> AxisData x

Modifier to position grid lines to line up with the ticks

axisGridAtBigTicks :: AxisData x -> AxisData x

Modifier to position grid lines to line up with only the major ticks

axisGridAtLabels :: AxisData x -> AxisData x

Modifier to position grid lines to line up with the labels

axisGridHide :: AxisData x -> AxisData x

Modifier to remove grid lines from an axis

axisLabelsOverride :: [(x, String)] -> AxisData x -> AxisData x

Modifier to change labels on an axis

axis_show_line :: Lens' AxisVisibility Bool

axis_show_ticks :: Lens' AxisVisibility Bool

axis_show_labels :: Lens' AxisVisibility Bool

axis_visibility :: forall x. Lens' (AxisData x) AxisVisibility

axis_viewport :: forall x. Lens' (AxisData x) (Range -> x -> Double)

axis_tropweiv :: forall x. Lens' (AxisData x) (Range -> Double -> x)

axis_ticks :: forall x. Lens' (AxisData x) [(x, Double)]

axis_labels :: forall x. Lens' (AxisData x) [[(x, String)]]

axis_grid :: forall x. Lens' (AxisData x) [x]

axis_line_style :: Lens' AxisStyle LineStyle

axis_label_style :: Lens' AxisStyle FontStyle

axis_grid_style :: Lens' AxisStyle LineStyle

axis_label_gap :: Lens' AxisStyle Double