Graphics.Rendering.Chart.Grid

Description

A container type for values that can be composed by horizonal and vertical layout.

Documentation

data Grid a

Abstract datatype representing a grid.

type Span = (Int, Int)

type SpaceWeight = (Double, Double)

When more space is available for an item than the total width of items, extra added space is proportional to 'space weight'.

tval :: a -> Grid a

A 1x1 grid from a given value, with no extra space.

tspan :: a -> Span -> Grid a

A WxH (measured in cells) grid from a given value, with space weight (1,1).

empty :: Grid a

A 1x1 empty grid.

nullt :: Grid a

A 0x0 empty grid.

(.|.) :: Grid a -> Grid a -> Grid a

A synonym for beside.

(./.) :: Grid a -> Grid a -> Grid a

A synonym for above.

above :: Grid a -> Grid a -> Grid a

aboveN :: [Grid a] -> Grid a

beside :: Grid a -> Grid a -> Grid a

besideN :: [Grid a] -> Grid a

overlay :: Grid a -> Grid a -> Grid a

One grid over the other. The first argument is shallow, the second is deep.

width :: Grid a -> Int

height :: Grid a -> Int

gridToRenderable :: Grid (Renderable a) -> Renderable a

weights :: SpaceWeight -> Grid a -> Grid a

Sets the space weight of *every* cell of the grid to given value.

aboveWide :: Grid a -> a -> Grid a

A value placed below the grid, occupying 1 row with the same horizontal span as the grid.

wideAbove :: a -> Grid a -> Grid a

A value occupying 1 row with the same horizontal span as the grid.

tallBeside :: a -> Grid a -> Grid a

A value placed to the left of the grid, occupying 1 column with the same vertical span as the grid.

besideTall :: Grid a -> a -> Grid a

A value placed to the right of the grid, occupying 1 column with the same vertical span as the grid.

fullOverlayUnder :: a -> Grid a -> Grid a

A value placed under a grid, with the same span as the grid.

fullOverlayOver :: a -> Grid a -> Grid a

A value placed over a grid, with the same span as the grid.