Graphics.Rendering.Chart.Legend

Description

Types and functions for handling the legend(s) on a chart. A legend is an area on the chart used to label the plotted values.

Documentation

data Legend x y

Constructors

Legend LegendStyle [(String, Rect -> BackendProgram ())]

data LegendStyle

Constructors

LegendStyle

Fields

_legend_label_style :: FontStyle

_legend_margin :: Double

_legend_plot_size :: Double

_legend_orientation :: LegendOrientation

_legend_position :: LegendPosition

data LegendOrientation

Legends can be constructed in two orientations: in rows (where we specify the maximum number of columns), and in columns (where we specify the maximum number of rows)

Constructors

LORows Int

LOCols Int

data LegendPosition

Defines the position of the legend, relative to the plot.

Constructors

LegendAbove

LegendBelow

LegendRight

LegendLeft

legendToRenderable :: Legend x y -> Renderable String

legend_label_style :: Lens' LegendStyle FontStyle

legend_margin :: Lens' LegendStyle Double

legend_plot_size :: Lens' LegendStyle Double

legend_orientation :: Lens' LegendStyle LegendOrientation

legend_position :: Lens' LegendStyle LegendPosition