Graphics.Rendering.Chart.Plot.Annotation

Description

Show textual annotations on a chart.

Documentation

data PlotAnnotation x y

Value for describing a series of text annotations to be placed at arbitrary points on the graph. Annotations can be rotated and styled.

Constructors

PlotAnnotation

Fields

_plot_annotation_hanchor :: HTextAnchor

_plot_annotation_vanchor :: VTextAnchor

_plot_annotation_angle :: Double

Angle, in degrees, to rotate the annotation about the anchor point.

_plot_annotation_style :: FontStyle

_plot_annotation_background :: Rectangle

Rectangle which style determines the background of the annotation text and which _rect_minsize determines the additional width and height of the background area

_plot_annotation_values :: [(x, y, String)]

plot_annotation_hanchor :: forall x y. Lens' (PlotAnnotation x y) HTextAnchor

plot_annotation_vanchor :: forall x y. Lens' (PlotAnnotation x y) VTextAnchor

plot_annotation_angle :: forall x y. Lens' (PlotAnnotation x y) Double

plot_annotation_style :: forall x y. Lens' (PlotAnnotation x y) FontStyle

plot_annotation_background :: forall x y. Lens' (PlotAnnotation x y) Rectangle

plot_annotation_values :: forall x y x y. Lens (PlotAnnotation x y) (PlotAnnotation x y) [(x, y, String)] [(x, y, String)]