Graphics.Rendering.Chart.Plot.AreaSpots

Description

Area spots are a collection of unconnected filled circles, with x,y position, and an independent z value to be represented by the relative area of the spots.

Documentation

data AreaSpots z x y

A collection of unconnected spots, with x,y position, and an independent z value to be represented by the area of the spot.

Constructors

AreaSpots

Fields

_area_spots_title :: String

_area_spots_linethick :: Double

_area_spots_linecolour :: AlphaColour Double

_area_spots_fillcolour :: Colour Double

_area_spots_opacity :: Double

_area_spots_max_radius :: Double

the largest size of spot

_area_spots_values :: [(x, y, z)]

area_spots_title :: forall z x y. Lens' (AreaSpots z x y) String

area_spots_linethick :: forall z x y. Lens' (AreaSpots z x y) Double

area_spots_linecolour :: forall z x y. Lens' (AreaSpots z x y) (AlphaColour Double)

area_spots_fillcolour :: forall z x y. Lens' (AreaSpots z x y) (Colour Double)

area_spots_opacity :: forall z x y. Lens' (AreaSpots z x y) Double

area_spots_max_radius :: forall z x y. Lens' (AreaSpots z x y) Double

area_spots_values :: forall z x y z x y. Lens (AreaSpots z x y) (AreaSpots z x y) [(x, y, z)] [(x, y, z)]

data AreaSpots4D z t x y

A collection of unconnected spots, with x,y position, an independent z value to be represented by the area of the spot, and in addition, a fourth variable t to be represented by a colour from a given palette. (A linear transfer function from t to palette is assumed.)

Constructors

AreaSpots4D

Fields

_area_spots_4d_title :: String

_area_spots_4d_linethick :: Double

_area_spots_4d_palette :: [Colour Double]

_area_spots_4d_opacity :: Double

_area_spots_4d_max_radius :: Double

the largest size of spot

_area_spots_4d_values :: [(x, y, z, t)]

area_spots_4d_title :: forall z t x y. Lens' (AreaSpots4D z t x y) String

area_spots_4d_linethick :: forall z t x y. Lens' (AreaSpots4D z t x y) Double

area_spots_4d_palette :: forall z t x y. Lens' (AreaSpots4D z t x y) [Colour Double]

area_spots_4d_opacity :: forall z t x y. Lens' (AreaSpots4D z t x y) Double

area_spots_4d_max_radius :: forall z t x y. Lens' (AreaSpots4D z t x y) Double

area_spots_4d_values :: forall z t x y z t x y. Lens (AreaSpots4D z t x y) (AreaSpots4D z t x y) [(x, y, z, t)] [(x, y, z, t)]