Graphics.Rendering.Chart.Plot.Bars

Description

Bar Charts

Documentation

data PlotBars x y

Constructors

PlotBars

Fields

_plot_bars_settings :: BarsSettings

_plot_bars_titles :: [String]

The title of each element of [y]. These will be shown in the legend.

_plot_bars_values_with_labels :: [(x, [(y, String)])]

The actual points to be plotted, and their labels

data PlotBarsStyle

Constructors

BarsStacked
Bars for a fixed x are stacked vertically on top of each other.

BarsClustered
Bars for a fixed x are put horizontally beside each other.

data PlotBarsSpacing

Constructors

BarsFixWidth Double
All bars have the same width in pixels.

BarsFixGap Double Double
(BarsFixGap g mw) means make the gaps between the bars equal to g, but with a minimum bar width of mw

data PlotBarsAlignment

How bars for a given (x,[y]) are aligned with respect to screen coordinate corresponding to x (deviceX).

Constructors

BarsLeft
The left edge of bars is at deviceX

BarsCentered
Bars are centered around deviceX

BarsRight
The right edge of bars is at deviceX

class PlotValue a => BarsPlotValue a where

Methods

barsIsNull :: a -> Bool

barsReference :: [a] -> a

The starting level for the chart, a function of some statistic (normally the lowest value or just const 0).

barsAdd :: a -> a -> a

data BarHorizAnchor

Constructors

BHA_Left

BHA_Centre

BHA_Right

data BarVertAnchor

Constructors

BVA_Bottom

BVA_Centre

BVA_Top

plotBars :: BarsPlotValue y => PlotBars x y -> Plot x y

plotHBars :: BarsPlotValue x => PlotBars y x -> Plot x y

plot_bars_style :: Lens' (PlotBars x y) PlotBarsStyle

plot_bars_item_styles :: Lens' (PlotBars x y) [(FillStyle, Maybe LineStyle)]

plot_bars_titles :: forall x y. Lens' (PlotBars x y) [String]

plot_bars_spacing :: Lens' (PlotBars x y) PlotBarsSpacing

plot_bars_alignment :: Lens' (PlotBars x y) PlotBarsAlignment

plot_bars_singleton_width :: Lens' (PlotBars x y) Double

plot_bars_label_bar_hanchor :: Lens' (PlotBars x y) BarHorizAnchor

plot_bars_label_bar_vanchor :: Lens' (PlotBars x y) BarVertAnchor

plot_bars_label_text_hanchor :: Lens' (PlotBars x y) HTextAnchor

plot_bars_label_text_vanchor :: Lens' (PlotBars x y) VTextAnchor

plot_bars_label_angle :: Lens' (PlotBars x y) Double

plot_bars_label_style :: Lens' (PlotBars x y) FontStyle

plot_bars_label_offset :: Lens' (PlotBars x y) Vector

plot_bars_values :: Lens' (PlotBars x y) [(x, [y])]

plot_bars_settings :: forall x y. Lens' (PlotBars x y) BarsSettings

plot_bars_values_with_labels :: forall x y x y. Lens (PlotBars x y) (PlotBars x y) [(x, [(y, String)])] [(x, [(y, String)])]

addLabels :: Show y => [(x, [y])] -> [(x, [(y, String)])]