Graphics.Rendering.Chart.Plot.Candle

Description

Candlestick charts for financial plotting

Documentation

data PlotCandle x y

Value defining a financial interval: opening and closing prices, with maxima and minima; and a style in which to render them. By convention, there are different fill styles depending on whether the price rises (open < close) or falls (close < open). (This plot type can also be re-purposed for statistical intervals, e.g. minimum, first quartile, median, third quartile, maximum.)

Constructors

PlotCandle

Fields

_plot_candle_title :: String

_plot_candle_line_style :: LineStyle

_plot_candle_fill :: Bool

_plot_candle_rise_fill_style :: FillStyle

_plot_candle_fall_fill_style :: FillStyle

_plot_candle_tick_length :: Double

_plot_candle_width :: Double

_plot_candle_centre :: Double

_plot_candle_values :: [Candle x y]

data Candle x y

A Value holding price intervals for a given x-coord. An alternative view is that these are statistical intervals: the 0th, 25th, 50th, 75th, and 100th percentiles.

Constructors

Candle

Fields

candle_x :: x

candle_low :: y

candle_open :: y

candle_mid :: y

candle_close :: y

candle_high :: y

plot_candle_title :: forall x y. Lens' (PlotCandle x y) String

plot_candle_line_style :: forall x y. Lens' (PlotCandle x y) LineStyle

plot_candle_tick_length :: forall x y. Lens' (PlotCandle x y) Double

plot_candle_width :: forall x y. Lens' (PlotCandle x y) Double

plot_candle_centre :: forall x y. Lens' (PlotCandle x y) Double

plot_candle_fill :: forall x y. Lens' (PlotCandle x y) Bool

plot_candle_rise_fill_style :: forall x y. Lens' (PlotCandle x y) FillStyle

plot_candle_fall_fill_style :: forall x y. Lens' (PlotCandle x y) FillStyle

plot_candle_values :: forall x y x y. Lens (PlotCandle x y) (PlotCandle x y) [Candle x y] [Candle x y]