Graphics.Rendering.Chart.Plot.ErrBars

Description

Plot series of points with associated error bars.

Documentation

data PlotErrBars x y

Value defining a series of error intervals, and a style in which to render them.

Constructors

PlotErrBars

Fields

_plot_errbars_title :: String

_plot_errbars_line_style :: LineStyle

_plot_errbars_tick_length :: Double

_plot_errbars_overhang :: Double

_plot_errbars_values :: [ErrPoint x y]

data ErrPoint x y

Constructors

ErrPoint

Fields

ep_x :: ErrValue x

ep_y :: ErrValue y

data ErrValue x

Value for holding a point with associated error bounds for each axis.

Constructors

ErrValue

Fields

ev_low :: x

ev_best :: x

ev_high :: x

symErrPoint :: (Num a, Num b) => a -> b -> a -> b -> ErrPoint a b

When the error is symmetric, we can simply pass in dx for the error.

Accessors

These accessors are generated by template haskell

plot_errbars_title :: forall x y. Lens' (PlotErrBars x y) String

plot_errbars_line_style :: forall x y. Lens' (PlotErrBars x y) LineStyle

plot_errbars_tick_length :: forall x y. Lens' (PlotErrBars x y) Double

plot_errbars_overhang :: forall x y. Lens' (PlotErrBars x y) Double

plot_errbars_values :: forall x y x y. Lens (PlotErrBars x y) (PlotErrBars x y) [ErrPoint x y] [ErrPoint x y]