Graphics.Rendering.Chart.Plot.FillBetween

Description

Plots that fill the area between two lines.

Documentation

data PlotFillBetween x y

Value specifying a plot filling the area between two sets of Y coordinates, given common X coordinates.

Constructors

PlotFillBetween

Fields

_plot_fillbetween_title :: String

_plot_fillbetween_style :: FillStyle

_plot_fillbetween_line :: Maybe LineStyle

_plot_fillbetween_values :: [(x, (y, y))]

Accessors

These accessors are generated by template haskell

plot_fillbetween_title :: forall x y. Lens' (PlotFillBetween x y) String

plot_fillbetween_style :: forall x y. Lens' (PlotFillBetween x y) FillStyle

plot_fillbetween_line :: forall x y. Lens' (PlotFillBetween x y) (Maybe LineStyle)

plot_fillbetween_values :: forall x y x y. Lens (PlotFillBetween x y) (PlotFillBetween x y) [(x, (y, y))] [(x, (y, y))]