Graphics.Rendering.Chart.Plot.Hidden

Description

Plots that don't show, but occupy space so as to effect axis scaling

Documentation

data PlotHidden x y

Value defining some hidden x and y values. The values are not displayed, but they still affect axis scaling.

Constructors

PlotHidden

Fields

_plot_hidden_x_values :: [x]

_plot_hidden_y_values :: [y]

plot_hidden_x_values :: forall x y x. Lens (PlotHidden x y) (PlotHidden x y) [x] [x]

plot_hidden_y_values :: forall x y y. Lens (PlotHidden x y) (PlotHidden x y) [y] [y]