Graphics.Rendering.Chart.Plot.Lines

Description

Line plots

Documentation

data PlotLines x y

Value defining a series of (possibly disjointed) lines, and a style in which to render them.

Constructors

PlotLines

Fields

_plot_lines_title :: String

_plot_lines_style :: LineStyle

_plot_lines_values :: [[(x, y)]]

The lines to be plotted

_plot_lines_limit_values :: [[(Limit x, Limit y)]]

Additional lines to be plotted, specified using the Limit type to allow referencing the edges of the plot area.

defaultPlotLineStyle :: LineStyle

hlinePlot :: String -> LineStyle -> b -> Plot a b

Helper function to plot a single horizontal line.

vlinePlot :: String -> LineStyle -> a -> Plot a b

Helper function to plot a single vertical line.

plot_lines_title :: forall x y. Lens' (PlotLines x y) String

plot_lines_style :: forall x y. Lens' (PlotLines x y) LineStyle

plot_lines_values :: forall x y. Lens' (PlotLines x y) [[(x, y)]]

plot_lines_limit_values :: forall x y. Lens' (PlotLines x y) [[(Limit x, Limit y)]]