Graphics.Rendering.Chart.Plot.Points

Description

Functions to plot sets of points, marked in various styles.

Documentation

data PlotPoints x y

Value defining a series of datapoints, and a style in which to render them.

Constructors

PlotPoints

Fields

_plot_points_title :: String

_plot_points_style :: PointStyle

_plot_points_values :: [(x, y)]

Accessors

These accessors are generated by template haskell

plot_points_title :: forall x y. Lens' (PlotPoints x y) String

plot_points_style :: forall x y. Lens' (PlotPoints x y) PointStyle

plot_points_values :: forall x y x y. Lens (PlotPoints x y) (PlotPoints x y) [(x, y)] [(x, y)]