Graphics.Rendering.Chart.Plot.Types

Description

Datatypes and functions common to the implementation of the various plot types.

Documentation

data Plot x y

Interface to control plotting on a 2D area.

Constructors

Plot

Fields

_plot_render :: PointMapFn x y -> BackendProgram ()

Given the mapping between model space coordinates and device coordinates, render this plot into a chart.

_plot_legend :: [(String, Rect -> BackendProgram ())]

Details for how to show this plot in a legend. For each item the string is the text to show, and the function renders a graphical sample of the plot.

_plot_all_points :: ([x], [y])

All of the model space coordinates to be plotted. These are used to autoscale the axes where necessary.

joinPlot :: Plot x y -> Plot x y -> Plot x y

Join any two plots together (they will share a legend).

class ToPlot a where

A type class abstracting the conversion of a value to a Plot.

Methods

toPlot :: a x y -> Plot x y

mapXY :: PointMapFn x y -> (x, y) -> Point

plot_render :: forall x y. Lens' (Plot x y) (PointMapFn x y -> BackendProgram ())

plot_legend :: forall x y. Lens' (Plot x y) [(String, Rect -> BackendProgram ())]

plot_all_points :: forall x y. Lens' (Plot x y) ([x], [y])