Graphics.Rendering.Chart.Plot.Vectors

Description

Vector plots

Documentation

data PlotVectors x y

Constructors

PlotVectors

Fields

_plot_vectors_title :: String

_plot_vectors_style :: VectorStyle

_plot_vectors_scale :: Double

Set to 1 (default) to normalize the length of vectors to a space between them (so that the vectors never overlap on the graph). Set to 0 to disable any scaling. Values in between 0 and 1 are also permitted to adjust scaling.

_plot_vectors_grid :: [(x, y)]

Provide a square-tiled regular grid.

_plot_vectors_mapf :: (x, y) -> (x, y)

Provide a vector field (R^2 -> R^2) function.

_plot_vectors_values :: [((x, y), (x, y))]

Provide a prepared list of (start,vector) pairs.

data VectorStyle

Constructors

VectorStyle

Fields

_vector_line_style :: LineStyle

_vector_head_style :: PointStyle

plotVectorField :: (PlotValue x, PlotValue y) => PlotVectors x y -> Plot x y

plot_vectors_mapf :: forall x y. Lens' (PlotVectors x y) ((x, y) -> (x, y))

plot_vectors_grid :: forall x y. Lens' (PlotVectors x y) [(x, y)]

plot_vectors_title :: forall x y. Lens' (PlotVectors x y) String

plot_vectors_style :: forall x y. Lens' (PlotVectors x y) VectorStyle

plot_vectors_scale :: forall x y. Lens' (PlotVectors x y) Double

plot_vectors_values :: forall x y. Lens' (PlotVectors x y) [((x, y), (x, y))]

vector_line_style :: Lens' VectorStyle LineStyle

vector_head_style :: Lens' VectorStyle PointStyle