Graphics.Rendering.Chart.SparkLine

Description

Sparklines are mini graphs inspired by Edward Tufte; see http://www.edwardtufte.com/bboard/q-and-a-fetch-msg?msg_id=0001OR and http://en.wikipedia.org/wiki/Sparkline for more information.

The original implementation (by Hitesh Jasani) used the gd package as a backend renderer, and is still available at http://hackage.haskell.org/package/hsparklines.

The present version integrates with the Chart package, in the sense that Sparklines are just another kind of (ToRenderable a => a), so they can be composed into grids and used with the rest of Chart.

dp :: [Double]
dp = [24,21,32.3,24,15,34,43,55,57,72,74,75,73,72,55,44]

sl = SparkLine barSpark dp
fopts = FileOptions (sparkSize sl) PNG
renderableToFile fopts (sparkLineToRenderable sl) "bar_spark.png" 

SparkLine type

data SparkLine

A sparkline is a single sequence of data values, treated as y-values. The x-values are anonymous and implicit in the sequence.

Constructors

SparkLine

Fields

sl_options :: SparkOptions

sl_data :: [Double]

Drawing options

data SparkOptions

Options to render the sparklines in different ways.

Constructors

SparkOptions

Fields

so_smooth :: Bool

smooth or bars

so_step :: Int

step size

so_height :: Int

graph height (pixels)

so_limits :: (Double, Double)

data point limits

so_bgColor :: Colour Double

background color

so_minColor :: Colour Double

color of minimum datapoint

so_maxColor :: Colour Double

color of maximum datapoint

so_lastColor :: Colour Double

color of last datapoint

so_minMarker :: Bool

display minimum marker

so_maxMarker :: Bool

display maximum marker

so_lastMarker :: Bool

display last marker

smoothSpark :: SparkOptions

Default options for a smooth sparkline.

barSpark :: SparkOptions

Default options for a barchart sparkline.

Size calculation

sparkSize :: SparkLine -> (Int, Int)

Return the width and height of the SparkLine.

Rendering function

renderSparkLine :: SparkLine -> BackendProgram (PickFn ())

Render a SparkLine to a drawing surface.

sparkLineToRenderable :: SparkLine -> Renderable ()

Create a renderable from a SparkLine.

sparkWidth :: SparkLine -> Int

Compute the width of a SparkLine, for rendering purposes.